platform/x86: Introduce an interface to Platform Services Hub
Introduce an interface driver for Platform Services Hub (PSH)
which can be found in Intel Tangier SoC.

PSH is a MinuteIA microarchitecture which shares some peripheral
with System Control Unit (SCU).

PSH has quite limited resources, in particularly memory (384KiB)
and runs on low frequency (100MHz). Due to the constrains it runs
modified version of Viper RTOS.

SCU controls it and here is the driver for that particular type of
communication.

Those three sysfs interfaces are as:
/sys/devices/platform/intel_mcu/control		to load the PSH payload
/sys/devices/platform/intel_mcu/fw_version	to read out version of payload
/sys/devices/platform/intel_mcu/log_level	to set and read payload's log level

Signed-off-by: Fang Hui <huix.fang@intel.com>
Signed-off-by: Lei Wen <lei.wen@intel.com>
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>